backend/requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
